// src/app/modules/admin/secretaria/estudante/model/checklist.model.ts
export interface ChecklistDocumentacao {
    rgEstudante: boolean;
    rgResponsavel: boolean;
    cpfEstudante: boolean;
    cpfResponsavel: boolean;
    comprovanteResidencia: boolean;
    declaracaoEscolar: boolean;
    certidaoNascimento: boolean;
    termoAudioVideo: boolean;
    atestadoMedico: boolean;
    historicoEscolar: boolean;

}